+++ src/phonetics/soundex.js
@@ -23,6 +23,13 @@ const TRANSLATIONS = {};
 
 LETTERS.forEach((letter, i) => TRANSLATIONS[letter] = CODES[i]);
 
+const REFINED_LETTERS = 'AEIOUYWHBPFVCKSGJQXZDTLMNR'.split(''),
+      REFINED_CODES = '000000DD112233344555667889'.split('');
+
+const REFINED_TRANSLATIONS = {};
+
+REFINED_LETTERS.forEach((letter, i) => REFINED_TRANSLATIONS[letter] = REFINED_CODES[i]);
+
 /**
  * Helpers.
  */
@@ -34,7 +41,7 @@ function pad(code) {
  * Function taking a single name and computing its Soundex code.
  *
  * @param  {string}  name - The name to process.
- * @return {string}       - The soundex code.
+ * @return {string}       - The Soundex code.
  *
  * @throws {Error} The function expects the name to be a string.
  */
@@ -56,7 +63,7 @@ export default function soundex(name) {
       tail += TRANSLATIONS[name[i]];
   }
 
-  // Encoding the tail
+  // Dropping first code's letter if duplicate
   if (tail.charAt(0) === TRANSLATIONS[firstLetter])
     tail = tail.slice(1);
 
@@ -64,4 +71,36 @@ export default function soundex(name) {
   const code = squeeze(tail).replace(/0/g, '');
 
   return pad(firstLetter + code);
+}
+
+/**
+ * Function taking a single name and computing its refined Soundex code.
+ *
+ * @param  {string}  name - The name to process.
+ * @return {string}       - The refined Soundex code.
+ *
+ * @throws {Error} The function expects the name to be a string.
+ */
+export function refined(name) {
+  if (typeof name !== 'string')
+    throw Error('talisman/phonetics/soundex:refined: the given name is not a string.');
+
+  name = deburr(name)
+    .toUpperCase()
+    .replace(/[^A-Z]/g, '');
+
+  const firstLetter = name.charAt(0);
+
+  // Process the code for the name's tail
+  let tail = '';
+
+  for (let i = 0, l = name.length; i < l; i++) {
+    if (REFINED_TRANSLATIONS[name[i]] !== 'D')
+      tail += REFINED_TRANSLATIONS[name[i]];
+  }
+
+  // Composing the code from the tail
+  const code = squeeze(tail);
+
+  return firstLetter + code;
 }